feat[DevTools]: Use Chrome DevTools Performance extension API
This change is a proof of concept of how the new Chrome DevTools
Performance extension API (https://bit.ly/rpp-e11y) can be used to
surface React runtime data directly in the Chrome DevTools Performance
panel.

To do this, the hooks in profilingHooks.js that mark beginning and end
of React measurements using [Performance marks](https://developer.mozilla.org/en-US/docs/Web/API/PerformanceMeasure)
                                            are modified to also use [Performance measure](https://developer.mozilla.org/en-US/docs/Web/API/PerformanceMeasure).
                with the `detail` field format specification of the
Performance extension API.

Because the marks in these hooks marks are used by React Profiler, they
are kept untouched and the calls to performance.measure are added on top
to surface them to the Chrome DevTools Performance panel, along with the
browser's native runtime data.

Because this is a proof of concept, not all the tasks and marks taken by
the React Profiler are added to the Chrome DevTools Performance panel
(f.e. update scheduling marks), but this could be done as a follow up of
this commit.

Note: to enable the user timings to be collected in the first place,
the React DevTools extension needs to be installed.